+++ inst/benchmark.R
@@ -48,10 +48,12 @@ benchmark <- function(path,
           n_trees = n_trees,
           max_iterations = max_iters,
           tree_threshold = threshold,
-          method = "largeVis-efficient-one",
+          method = "largeVis",
           tree_type = "",
           searchtype = "",
-          eps = 0)
+          eps = 0,
+          K = K,
+          machine = "aws")
         print(one_result)
         readr::write_csv(one_result, path = "results.csv", append = TRUE)
         if (precision == K) break
@@ -179,46 +181,89 @@ benchmarkAnnoy <- function(path,
       method = method,
       tree_type = "",
       searchtype = "",
-      eps = 0)
+      eps = 0,
+      K = K,
+      machine = "aws")
     print(one_result)
     readr::write_csv(one_result, path = "results.csv", append = TRUE)
   }
 }
 
 require( largeVis )
-path <- "/mnt/hfsshare/DATASETS/sift/siftknns.txt"
+path <- "./siftknns.txt"
 samplepath <- "./samples.Rda"
 
-# Annoyresults <- benchmarkAnnoy(path,
-#                                samplepath,
-#                                tree_range = c(10, 20, 50, 100, 200, 400),
-#                                n = 10000,
-#                                K = 100,
-#                                full = FALSE)
-# results <- benchmark(path,
-#                      samplepath,
-#                      n = 10000,
-#                      tree_range = c(10, 20, 50, 100, 200),
-#                      thresholds = c(128),
-#                      iters = c(1, 0, 2, 3),
-#                      K = 100)
-results2 <- benchmark(path,
+Annoyresults <- benchmarkAnnoy(path,
+                               samplepath,
+                               tree_range = c(10, 20, 50, 100, 200, 400),
+                               n = 10000,
+                               K = 50,
+                               full = TRUE)
+results <- benchmark(path,
                      samplepath,
                      n = 10000,
-                     tree_range = c(10, 20, 50),
-                     thresholds = c(10, 20, 50, 80, 256, 512),
-                     iters = c(1),
-                     K = 100)
-# Annoyresults2 <- benchmarkAnnoy(path,
-#                                samplepath,
-#                                tree_range = c(10, 20, 50, 100, 200, 400),
-#                                n = 10000,
-#                                K = 100,
-#                                full = TRUE)
+                     tree_range = c(10, 20, 50, 100, 200),
+                     thresholds = c(128),
+                     iters = c(1, 0, 2, 3),
+                     K = 50)
+results2 <- benchmark(path,
+                    samplepath,
+                    n = 10000,
+                    tree_range = c(10, 20, 50),
+                    thresholds = c(10, 20, 50, 80, 256, 512),
+                    iters = c(1),
+                    K = 50)
+results2 <- benchmark(path,
+                      samplepath,
+                      n = 10000,
+                      tree_range = c(10),
+                      thresholds = c(200, 400, 800),
+                      iters = c(0,1,2),
+                      K = 50)
+results2 <- benchmark(path,
+                      samplepath,
+                      n = 10000,
+                      tree_range = c(20),
+                      thresholds = c(100, 200, 400),
+                      iters = c(0,1,2),
+                      K = 50)
+results2 <- benchmark(path,
+                      samplepath,
+                      n = 10000,
+                      tree_range = c(40),
+                      thresholds = c(50, 100, 200),
+                      iters = c(0,1,2),
+                      K = 50)
+results2 <- benchmark(path,
+                      samplepath,
+                      n = 10000,
+                      tree_range = c(80),
+                      thresholds = c(25, 50, 100),
+                      iters = c(0,1,2),
+                      K = 50)
 # RANNresults <- benchmarkRANN(path,
 #                              samplepath,
 #                              epss = c(.1, .5, 1,2,5),
 #                              n = 10000,
-#                              K = 100)
-
-print(results)
+#                              K = 500)
+results2 <- benchmark(path,
+                     samplepath,
+                     n = 10000,
+                     tree_range = c(2),
+                     thresholds = c(50, 100, 250),
+                     iters = c(0,1,2),
+                     K = 50)
+results2 <- benchmark(path,
+                     samplepath,
+                     n = 10000,
+                     tree_range = c(4),
+                     thresholds = c(125),
+                     iters = c(0,1,2),
+                     K = 50)
+results2 <- benchmark(path,
+                     samplepath,
+                     n = 10000,
+                     tree_range = c(5),
+                     thresholds = c(20, 40, 100),
+                     iters = c(0,1,2),
+                     K = 50)