+++ tests/testthat/testdistance.R
@@ -2,7 +2,6 @@ context("distance")
 
 test_that("Euclidean distances are correct", {
   set.seed(1974)
-  RcppArmadillo::armadillo_set_seed(1974)
   test_matrix <- matrix(rnorm(100), nrow = 10)
   distances <- as.matrix(dist(test_matrix, method = "euclidean"))
   index_matrix <- matrix(c(rep(0:9, each = 10), rep(0:9, 10)),
@@ -20,7 +19,6 @@ test_that("Euclidean distances are correct", {
 
 test_that("Cosine distances are correct", {
   set.seed(1974)
-  RcppArmadillo::armadillo_set_seed(1974)
   cos.sim <- function(x, i, j) {
     A <- x[, i]
     B <- x[, j]